Refactor component structure and move files to new directories
