+++ frontend/src/components/user_stats/GameHistory.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Card from "./ui/Card";
+import Card from "./../ui/Card";
 
 interface GameHistoryItem {
 	id: number;
